\name{randomizeBE-package}
\alias{randomizeBE-package}
\alias{randomizeBE}
\docType{package}
\title{
Create a randomisation list for cross-over studies.
  }
\description{
Randomises subjects, patients in groups of (treatment) sequences.\cr
If a blocksize is given, the randomisation will be done within blocks.\cr
If more then one blocksizes are given, the size of blocks will be chosen randomly
from the blocksizes.\cr
The randomisation may be controlled by a Wald-Wolfowitz runs test.
  }
\details{
\tabular{ll}{
Package: \tab randomizeBE\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2012-07-05\cr
License: \tab GPL (>=2)\cr
  }
}
\author{
D. Labes\cr
with part of the code for using the runs test for randomness \cr
according to a suggestion by Helmut Schuetz.

Maintainer: D.Labes <detlewlabes@gmx.de>
}

